exports types dir in skd and changes imports in manager
