+++ packages/linode-js-sdk/src/managed/index.ts
@@ -1,2 +1,3 @@
 export * from './managed';
 export * from './managed.schema';
+export * from './types';
