+++ packages/manager/src/features/Managed/Monitors/MonitorTableContent.tsx
@@ -1,4 +1,4 @@
-import { ManagedServiceMonitor } from 'linode-js-sdk/lib/managed/types';
+import { ManagedServiceMonitor } from 'linode-js-sdk/lib/managed';
 import * as React from 'react';
 
 import TableRowEmpty from 'src/components/TableRowEmptyState';
