+++ packages/manager/src/features/Managed/Monitors/monitorMaps.ts
@@ -1,4 +1,4 @@
-import { MonitorStatus } from 'linode-js-sdk/lib/managed/types';
+import { MonitorStatus } from 'linode-js-sdk/lib/managed';
 import Disabled from 'src/assets/icons/monitor-disabled.svg';
 import Bad from 'src/assets/icons/monitor-failed.svg';
 import Good from 'src/assets/icons/monitor-ok.svg';
